+++ src/include/fdw.h
@@ -245,7 +245,8 @@ bool		chfdw_pushdown_regex_ok(void);
 extern void
 			chfdw_extract_options(List * defelems, char **driver, char **host, int *port,
 								  char **dbname, char **username, char **password,
-								  char **compression, tls_mode * tls);
+								  char **compression, tls_mode * tls,
+								  tls_version * min_tls_version);
 extern List * chfdw_parse_options(const char *options, bool with_comma, bool with_equal);
 
 /* in deparse.c */